+++ mcww_web/js/misc/layout.js
@@ -57,10 +57,34 @@ onUiLoaded(() => {
     window.matchMedia('(display-mode: standalone)').addEventListener('change', checkForIsInsidePWA);
 });
 
-
+const GRADIO_APP_BROKEN_MESSAGE = "Connection to the server was lost. Attempting reconnection...";
 const BLACKLISTED_TOASTED_MESSAGES = [
     "Waiting for file(s) to finish uploading, please retry.",
+    "Ожидание завершения загрузки файла(ов), пожалуйста, повторите попытку.",
+    "Laukiama, kol bus baigti įkelti failai, prašome bandyti dar kartą.",
+    "ファイルのアップロードが完了するのを待っています。再度お試しください",
+    "等待文件上传完成，请稍后重试。",
+    "等待檔案上傳完成，請稍後重試。",
+    "Чекаємо завершення завантаження файлу(ів), будь ласка, спробуйте ще раз.",
+    "Aguardando o(s) arquivo(s) terminar(em) de ser carregado(s), por favor, tente novamente.",
+    "Oczekiwanie na zakończenie przesyłania pliku(ów), proszę spróbować ponownie.",
+    "Wachten tot het uploaden van bestand(en) is voltooid, probeer het opnieuw.",
+    "파일 업로드가 완료될 때까지 기다렸다가 다시 시도해주세요.",
+    "En attente de la fin du téléchargement des fichier(s), veuillez réessayer.",
+    "Odotetaan tiedostojen lataamisen päättymistä, yritä uudelleen.",
+    "Warten auf das Hochladen der Datei(en), bitte versuchen Sie es erneut.",
+    "Väntar på att fil(er) ska vara klara att ladda upp, vänligen försök igen.",
+    "Fayl(lar) yuklanib bo'lishini kutish, iltimos, qayta urinib ko'ring.",
+    "கோப்பு(கள்) பதிவேற்றம் முடிவடைய காத்திருக்கிறது, தயவுசெய்து மீண்டும் முயற்சிக்கவும்.",
+    "Esperando a que se complete(n) la(s) carga(s) de archivo(s), por favor inténtelo de nuevo.",
+    "إنتظر انتهاء رفع الملف(ات)، يرجى إعادة المحاولة.",
+    "फाइलें अपलोड होने का इंतजार कर रहे हैं, कृपया पुन: प्रयास करें।",
+    "מחכים לסיום העלאה של הקבצים, אנא נסה שוב.",
+    "در انتظار به پایان رسیدن بارگذاری فایل‌ها، لطفاً مجدداً امتحان کنید.",
+
     "Video not playable",
+    GRADIO_APP_BROKEN_MESSAGE,
+    "Connection re-established.",
 ];
 
 onUiUpdate((updatedElements) => {
@@ -69,6 +93,9 @@ onUiUpdate((updatedElements) => {
         const text = toastMessage.querySelector(".toast-text").textContent;
         if (BLACKLISTED_TOASTED_MESSAGES.includes(text)) {
             toastMessage.style.display = "none";
+        }
+        if (text === GRADIO_APP_BROKEN_MESSAGE) {
+            onGradioAppBroken();
         }
     });
 });